"fixed" LocalSocketCommChannel inputstream bug
PrebufferedInputstream gave garbled data when reading with DBus. Added
BufferedInputStream inside the PreBufferedInputStream class, seems to
fix the problem.

+++ extensions/localsocket/src/jolie/net/LocalSocketListener.java
@@ -63,7 +63,7 @@ public class LocalSocketListener extends CommListener
 	@Override
 	public void shutdown()
 	{
-		if ( !socketAddress.isAbstract() ) {
+		if ( !socketAddress.isAbstract() && !inputPort().messageBus()) {
 			new File( socketAddress.getPath() ).delete();
 		}
 	}
